+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+# [0.2.4] - 2025-07-17
+
+### Added
+
+- Corporation Overview
+
+### Fixed
+
+- API schema Issue with 403 errors
+
 # [0.2.3] - 2025-07-11
 
 ## Added